# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: CARMA View API
  description: The core's API for the UI and the MCP server (docs/carma-spec.md, «Контракт 4»). Symbol IDs contain spaces, '/' and '#', so they always travel as the query parameter id.
  version: '0.3'
paths:
...
  /api/v0/app/state:
    get:
      summary: State
      operationId: state_api_v0_app_state_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationState'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v0/app/pick-folder:
    post:
      summary: Pick Folder
      operationId: pick_folder_api_v0_app_pick_folder_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FolderChoice'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v0/app/inspect:
    post:
      summary: Inspect
      operationId: inspect_api_v0_app_inspect_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectPath'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectInspection'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v0/app/open:
    post:
      summary: Open Project
      operationId: open_project_api_v0_app_open_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenProject'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationState'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v0/app/reindex:
    post:
      summary: Reindex
      operationId: reindex_api_v0_app_reindex_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationState'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v0/app/cancel:
    post:
      summary: Cancel
      operationId: cancel_api_v0_app_cancel_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationState'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v0/app/close:
    post:
      summary: Close
      operationId: close_api_v0_app_close_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationState'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v0/app/forget:
    post:
      summary: Forget
      operationId: forget_api_v0_app_forget_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecentProject'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationState'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ApplicationState:
      properties:
        managed:
          type: boolean
          title: Managed
        token:
          anyOf:
            - type: string
            - type: 'null'
          title: Token
        active:
          anyOf:
            - $ref: '#/components/schemas/Project'
            - type: 'null'
        projects:
          items:
            $ref: '#/components/schemas/Project'
          type: array
          title: Projects
          default: []
        job:
          anyOf:
            - $ref: '#/components/schemas/ProjectJob'
            - type: 'null'
        warnings:
          items:
            type: string
          type: array
          title: Warnings
          default: []
      additionalProperties: false
      type: object
      required:
        - managed
      title: ApplicationState
    BuildChoice:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        kind:
          type: string
          enum:
            - saved
            - commands
            - build
            - preset
            - cmake
          title: Kind
        path:
          type: string
          title: Path
        configuration:
          anyOf:
            - type: string
            - type: 'null'
          title: Configuration
      additionalProperties: false
      type: object
      required:
        - id
        - name
        - kind
        - path
      title: BuildChoice
...
    ErrorBody:
      properties:
        code:
          type: string
          title: Code
          description: not_found, validation_failed or another machine-readable code
        message:
          type: string
          title: Message
        details:
          items: {}
          type: array
          title: Details
          default: []
      additionalProperties: false
      type: object
      required:
        - code
        - message
      title: ErrorBody
    ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
      additionalProperties: false
      type: object
      required:
        - error
      title: ErrorResponse
...
    FolderChoice:
      properties:
        path:
          anyOf:
            - type: string
            - type: 'null'
          title: Path
      additionalProperties: false
      type: object
      required:
        - path
      title: FolderChoice
...
    OpenProject:
      properties:
        path:
          type: string
          minLength: 1
          title: Path
        build_id:
          anyOf:
            - type: string
            - type: 'null'
          title: Build Id
        replace_id:
          anyOf:
            - type: string
            - type: 'null'
          title: Replace Id
          description: Relocated recent project to replace after successful opening
      additionalProperties: false
      type: object
      required:
        - path
      title: OpenProject
...
    Project:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        path:
          type: string
          title: Path
        available:
          type: boolean
          title: Available
        ready:
          type: boolean
          title: Ready
        opened_at:
          anyOf:
            - type: string
            - type: 'null'
          title: Opened At
      additionalProperties: false
      type: object
      required:
        - id
        - name
        - path
        - available
        - ready
      title: Project
    ProjectInspection:
      properties:
        project:
          $ref: '#/components/schemas/Project'
        choices:
          items:
            $ref: '#/components/schemas/BuildChoice'
          type: array
          title: Choices
      additionalProperties: false
      type: object
      required:
        - project
        - choices
      title: ProjectInspection
    ProjectJob:
      properties:
        id:
          type: string
          title: Id
        path:
          type: string
          title: Path
        status:
          type: string
          enum:
            - running
            - ready
            - failed
            - cancelled
          title: Status
        stage:
          type: string
          title: Stage
        message:
          type: string
          title: Message
        done:
          type: integer
          title: Done
        total:
          type: integer
          title: Total
        error:
          anyOf:
            - type: string
            - type: 'null'
          title: Error
      additionalProperties: false
      type: object
      required:
        - id
        - path
        - status
        - stage
        - message
        - done
        - total
      title: ProjectJob
    ProjectPath:
      properties:
        path:
          type: string
          minLength: 1
          title: Path
      additionalProperties: false
      type: object
      required:
        - path
      title: ProjectPath
    RecentProject:
      properties:
        id:
          type: string
          title: Id
      additionalProperties: false
      type: object
      required:
        - id
      title: RecentProject
...